+++ PinterestClone/src/main/java/com/sparta/pinterestclone/domain/pin/ResponseDto/CommentResponseDto.java
@@ -1,10 +1,12 @@
-package com.sparta.pinterestclone.domain.pin.dto;
+package com.sparta.pinterestclone.domain.pin.ResponseDto;
 
 import com.sparta.pinterestclone.domain.pin.entity.Comment;
 import lombok.Builder;
+import lombok.Getter;
 
 import java.time.LocalDateTime;
 
+@Getter
 public class CommentResponseDto {
     private String nickname;
     private String comment;
@@ -12,7 +14,7 @@ public class CommentResponseDto {
     private LocalDateTime modifiedAt;
 
     @Builder
-    public CommentResponseDto (Comment comment ){
+    public CommentResponseDto (Comment comment){
         nickname = comment.getUser().getNickname();
         this.comment = comment.getComment();
         createdAt = comment.getCreatedAt();
